Project card done with modif style
