+++ mobile/navigation/index.tsx
@@ -20,10 +20,10 @@ import { ColorSchemeName } from "react-native";
 import useColorScheme from "../hooks/useColorScheme";
 import ModalScreen from "../screens/ModalScreen";
 import NotFoundScreen from "../screens/NotFoundScreen";
-import TabOneScreen from "../screens/TabOneScreen";
-import TabTwoScreen from "../screens/TabTwoScreen";
-import TabThreeScreen from "../screens/TabThreeScreen";
-import TabFourScreen from "../screens/TabThreeScreen";
+import HomeScreen from "../screens/HomeScreen";
+import TasksScreen from "../screens/TasksScreen";
+import ProjectsScreen from "../screens/ProjectsScreen";
+import TabFourScreen from "../screens/ProjectsScreen";
 import {
 	RootStackParamList,
 	RootTabParamList,
@@ -106,7 +106,7 @@ function BottomTabNavigator() {
 		>
 			<BottomTab.Screen
 				name="TabOne"
-				component={TabOneScreen}
+				component={HomeScreen}
 				options={{
 					title: "Home",
 					//   tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
@@ -114,7 +114,7 @@ function BottomTabNavigator() {
 			/>
 			<BottomTab.Screen
 				name="TabTwo"
-				component={TabTwoScreen}
+				component={TasksScreen}
 				options={{
 					title: "Tasks",
 					//   tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
@@ -122,7 +122,7 @@ function BottomTabNavigator() {
 			/>
 			<BottomTab.Screen
 				name="TabThree"
-				component={TabThreeScreen}
+				component={ProjectsScreen}
 				options={{
 					title: "Projects",
 					//   tabBarIcon: ({ color }) => <TabBarIcon name="code" color={color} />,
