+++ mobile/screens/TasksScreen.tsx
@@ -3,16 +3,16 @@ import { StyleSheet } from "react-native";
 import EditScreenInfo from "../components/EditScreenInfo";
 import { Text, View } from "../components/Themed";
 
-export default function TabThreeScreen() {
+export default function TasksScreen() {
 	return (
 		<View style={styles.container}>
-			<Text style={styles.title}>Tab Three</Text>
+			<Text style={styles.title}>Tab Two</Text>
 			<View
 				style={styles.separator}
 				lightColor="#eee"
 				darkColor="rgba(255,255,255,0.1)"
 			/>
-			<EditScreenInfo path="/screens/TabThreeScreen.tsx" />
+			<EditScreenInfo path="/screens/TasksScreen.tsx" />
 		</View>
 	);
 }
